ASoC: tegra-alt: Fix coverity defect
Avoids integer overflow, bad bit shift operation

Coverity ID 29615

Bug 200192147

Reviewed-on: http://git-master/r/1258003
GVS: Gerrit_Virtual_Submit

+++ sound/soc/tegra-alt/tegra_asoc_machine_alt.c
@@ -1950,7 +1950,7 @@ static int tegra_machine_get_format(u64 *p_formats, char *fmt)
 
 	for (i = 0; i < ARRAY_SIZE(bit_format); i++) {
 		if (strcmp(bit_format[i], fmt) == 0) {
-			*p_formats = 1 << i;
+			*p_formats = (u64)1 << i;
 			return 0;
 		}
 	}
